core/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.14.
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.4.0
psycopg2==2.9.5
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
